Integrated ProtocolStarter with the SyncStatusProvider
